Prevent wp_convert_widget_settings() from being erronously called
The Efficient_Multidimensional_Setting_Sanitizing functionality was storing a widget settings
array which lacked _multiwidget=1 and so when this value was supplied to the pre_option_widget_{id_base}
when WP_Widget::get_settings() was called, it would end up calling wp_convert_widget_settings() and destroying
the instance data.

This only applied when Widget Posts was disabled because the Widget_Settings ArrayIterator always has _multiwidget=1.